+++ src/Glob/GlobExtensions.cs
@@ -10,26 +10,17 @@ namespace GlobExpressions
     {
         public static IEnumerable<DirectoryInfo> GlobDirectories(this DirectoryInfo di, string pattern)
         {
-            var glob = new Glob(pattern, GlobOptions.Compiled);
-            var truncateLength = di.FullName.Length + 1;
-
-            return di.EnumerateDirectories("*", SearchOption.AllDirectories).Where(info => glob.IsMatch(info.FullName.Remove(0, truncateLength)));
+            return Glob.Directories(di, pattern);
         }
 
         public static IEnumerable<FileInfo> GlobFiles(this DirectoryInfo di, string pattern)
         {
-            var glob = new Glob(pattern, GlobOptions.Compiled);
-            var truncateLength = di.FullName.Length + 1;
-
-            return di.EnumerateFiles("*", SearchOption.AllDirectories).Where(info => glob.IsMatch(info.FullName.Remove(0, truncateLength)));
+            return Glob.Files(di, pattern);
         }
 
         public static IEnumerable<FileSystemInfo> GlobFileSystemInfos(this DirectoryInfo di, string pattern)
         {
-            var glob = new Glob(pattern, GlobOptions.Compiled);
-            var truncateLength = di.FullName.Length + 1;
-
-            return di.EnumerateFileSystemInfos("*", SearchOption.AllDirectories).Where(info => glob.IsMatch(info.FullName.Remove(0, truncateLength)));
+            return Glob.FilesAndDirectories(di, pattern);
         }
     }
 }